Fix imports for restructured test utilities
