+++ tests/characters/test_alice_can_grant_and_revoke.py
@@ -4,7 +4,7 @@ import maya
 from umbral.fragments import KFrag
 
 from nucypher.crypto.api import keccak_digest
-from nucypher.utilities.network import _TEST_KNOWN_URSULAS_CACHE
+from nucypher.utilities.sandbox import _TEST_KNOWN_URSULAS_CACHE
 
 
 class MockPolicyCreation:
