+++ tests/characters/test_ursula_prepares_to_act_as_mining_node.py
@@ -3,8 +3,7 @@ from eth_keys.datatypes import Signature as EthSignature
 
 from nucypher.characters import Ursula
 from nucypher.crypto.powers import SigningPower, CryptoPower
-from nucypher.utilities.blockchain import make_ursulas
-from nucypher.utilities.network import MockRestMiddleware
+from nucypher.utilities.sandbox import make_ursulas, MockRestMiddleware
 
 
 @pytest.mark.skip("To be implemented...?")
